arreglando persistencia

+++ src/main/java/com/guilder/hospital/repositories/DiagnosticRepository.java
@@ -4,5 +4,5 @@ package com.guilder.hospital.repositories;
 import com.guilder.hospital.models.Diagnostic;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DiagnosticRepository extends CrudRepository<Diagnostic,Double> {
+public interface DiagnosticRepository extends CrudRepository<Diagnostic,Long> {
 }
